android/requirements.txt
requests
pillow
kivy
kivymd